I2C Expansions: Add I2C mux and gpio expander
Initial python implementation of these two.
I have plans to re-write these in C later on, but for now these should
suffice in getting started and providing a (hopefully) concrete API

Added classes to wrap the I2C mux and GPIO expander.
These should be transparent and with the same interface as machine.(I2C|Pin)
The I2C interface will automatically select itself when used on the MUX.
Since the I2C mux is fairly transparent, it just calls the parent bus
functions most of the time with a simple wrapper to select the bus.