Improve consistency of event loop
The scheduler code works great on the badge, but it relies on a quirk of how the event loop is handled in micropython. Specifically, we instantiate an _EventBus and store the singleton at module scope, which is used by the Scheduler, but those are both instantiated before asyncio.run is called. That causes a new eventloop to be created. Micropython tolerates awaiting futures created before the most recent invocation of new_event_loop, but CPython doesn't. The upshot of which is that the scheduler crashes the simulator code. I've made a change to replace the module-global eventbus with a callable that returns a singleton, and moved some code into a codepath that doesn't execute until after asyncio.run is called.