Add 0xA0 to EEPROM address detection for M24C16-R
The docs list M24C16-RMN6TP as a supported EEPROM for hexpansions, but the M24C16-R series uses addresses 0xA0 to 0xA8 for its page addressing. The code currently only supports 0x50 and 0x57 for addresses.

Note: THIS COMMIT IS NOT TESTED. I'm providing this in the hopes of speeding up an OTA rollout so we can get the nullsector hexpansions programmed and running.