Scheduler performance improvement
Since the first thing every app does every frame is clear the
framebuffer, there seems to be no point in asking every foreground
application to render itself because only the top-most app in the stack
will overwrite everything rendered by all the other apps in the stack.

This change renders only the top-most app in the foreground apps stack,
completely eliminating the time spend rendering the launcher and more or
less halving the end_frame step because the ctx drawlists are much
shorter.

This should yield a noticable improvement in the framerate of graphical
tildagon apps