Add support for a beta channel
This improves the OTA app to parse version numbers better, with the following rules:

- Major, Minor and Patch versions are parsed as numbers rather than strings, so v1.10.0 > v1.9.0
- Development versions where there's a 'number of commits ahead' item are counted as ahead of the version they're on
- If there's an rc after the patch number, it's treated as ahead of the previous version, so v1.9.0 < v1.10.0rc1 < v1.10.0rc1-1-aaaaaa < v1.10.0 < v1.10.0-1-aaaaaa

If there is an rc in the tag name, it won't overwrite the `latest` release, it will overwrite the `preview` release.

The setting `update_channel` will work as normal if unset or set to `latest`. Set it to `preview` to get release candidates.