customisable backgrounds
Add 2 default backgrounds with the ability to add more in a backgrounds folder. backgrounds apps follow a standard app pattern, the update and draw functions are called by firmware apps to create the background. backgrounds can be shared with the app store with the category set to 'Background' and can be updated in the normal way.