Custom background (#210)
* customisable backgrounds

Add 2 default backgrounds with the ability to add more in a backgrounds folder. backgrounds apps follow a standard app pattern, the update and draw functions are called by firmware apps to create the background. backgrounds can be shared with the app store with the category set to 'Background' and can be updated in the normal way.

* handle race condition between update and draw functions when selecting category

* fix missed directory change

* add save and restore to prevent menu fonts being changed by apps

also added backgrounds to sponsors, power off and update apps